refactor(app/inbound): response body metrics submodule
